functions/requirements.txt
functions-framework==3.*
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
